Save reported question to response records. Add error popover
